+++ src/main/java/com/flint/flint/club/domain/comment/ClubCommentLike.java
@@ -1,9 +1,8 @@
 package com.flint.flint.club.domain.comment;
 
-import com.flint.flint.member.domain.Member;
+import com.flint.flint.member.domain.main.Member;
 import jakarta.persistence.*;
 import lombok.Builder;
-import lombok.Data;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
