+++ src/main/java/com/flint/flint/mail/controller/MailController.java
@@ -34,29 +34,16 @@ import static com.flint.flint.common.spec.ResultCode.USER_MANY_REQUEST;
 @RequestMapping("/api/v1/mail")
 public class MailController {
 
-    private final static Long Expiration = 1800000L; //한시간
 
-    private final RedisUtil redisUtil;
     private final MailService mailService;
-    private final MemberService memberService;
-    private final AuthenticationService authenticationService;
-    private final RateLimitService rateLimitService;
-
     /**
      * 인증번호 보내기
      * 검증: API호출 횟수
      */
     @PostMapping("/send")
-    public ResponseForm<?> sendEmail(@RequestBody SendEmailAuthNumberReqeust request, @AuthenticationPrincipal AuthorityMemberDTO authorityMemberDTO) {
+    public ResponseForm<EmailAuthNumberRespose> sendEmail(@RequestBody SendEmailAuthNumberReqeust request, @AuthenticationPrincipal AuthorityMemberDTO authorityMemberDTO) {
         Long key = authorityMemberDTO.getId();
-        if (rateLimitService.checkAPICall(key)) { //API호출 횟수 검증
-            int authNumber = mailService.sendCodeEmail(request.getEmail());
-            redisUtil.saveAuthNumber(key, String.valueOf(authNumber), Expiration);  //인증번호 검증을 위해 레디스에 저장
-            EmailAuthNumberRespose response = new EmailAuthNumberRespose(authNumber);
-            return new ResponseForm<>(response);
-        } else { // API호출 한 시간에 10 번 이상 호출 한 경우 에러코드 보냄
-            return new ResponseForm<>(USER_MANY_REQUEST);
-        }
+        return new ResponseForm<>(mailService.sendCodeEmail(request.getEmail(), key));
     }
 
     /**
@@ -65,13 +52,8 @@ public class MailController {
      * 검증: 이메일 인증코드
      */
     @PostMapping("/success/auth")
-    public ResponseForm<?> successUniversityAuth(@RequestBody VeriyEmailAuthnumberRequest request, @AuthenticationPrincipal AuthorityMemberDTO authorityMemberDTO) {
+    public ResponseForm<AuthenticationResponse> successUniversityAuth(@RequestBody VeriyEmailAuthnumberRequest request, @AuthenticationPrincipal AuthorityMemberDTO authorityMemberDTO) {
         Long key = authorityMemberDTO.getId();
-        if (!(String.valueOf(redisUtil.findEmailAuthNumberByKey(key)).equals(String.valueOf(request.getAuthNumber()))))  //레디스에 저장한 인증번호와 다르다면 에러코드 보냄
-            return new ResponseForm<>(MAIL_AUTHNUMBER_NOT);
-        Member member = memberService.getMember(authorityMemberDTO.getId());
-        member.updateAuthority(Authority.AUTHUSER);
-        AuthenticationResponse authenticationResponse = authenticationService.generateToken(member); //인증 한 유저권한을 담은 토큰 발급
-        return new ResponseForm<>(authenticationResponse);
+        return new ResponseForm<>(mailService.successEmailAuth(request, key));
     }
 }
